Externalize Tooltip component
